requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
awscli==1.19.10
Babel==2.9.0
bcrypt==3.2.0
botocore==1.20.10
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.3
cryptography==3.3.1
decorator==4.4.2
distlib==0.3.1
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
infinity==1.5
iniconfig==1.1.1
intervals==0.9.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
phonenumbers==8.12.15
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.4
PyYAML==5.3.1
rope==0.18.0
rsa==4.5
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utc==0.11.0
toml==0.10.2
urllib3==1.26.3
validators==0.18.1
virtualenv==20.2.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
WTForms-Components==0.10.4